Add European mode
For Aloodo, European mode works the same as regular
mode.

This is to help test browser extensions that look
for sites that treat European and non-European users
differently.

Extensions can test if the page they're on has Aloodo
and European mode is available, with:

```
window.aloodo && typeof(aloodo.setEuropeanMode)
```

Example (work in progress):
  https://github.com/dmarti/trans-europa-express